Firmware: Uebersetzungsfehler unter Arduino-Core 3.x beheben
Der CI-Build brach in Settings.cpp ab: esp_read_mac() und
ESP_MAC_WIFI_STA kommen seit Arduino-Core 3.x nicht mehr ueber
Arduino.h. Beim Durchsehen fielen drei weitere Stellen auf, die als
naechstes gescheitert waeren oder im Betrieb gestoert haetten:

  * esp_reset_reason() braucht ebenfalls einen eigenen Include
    (esp_system.h)
  * NimBLE 2.x nimmt die Sendeleistung in dBm statt als
    esp_power_level_t; ESP_PWR_LVL_P9 existiert dort nicht mehr
  * Wire.requestFrom() ist auf int und uint8_t ueberladen - ohne Cast
    ist der Aufruf mehrdeutig
  * ledcAttach() lief bei jeder Helligkeitsaenderung erneut; das
    laesst die Hintergrundbeleuchtung flackern. Jetzt einmalig in
    Screen::begin(), setBrightness() schreibt nur noch den Tastgrad.

Ausserdem eine ungenutzte Farbkonstante entfernt.

+++ firmware/src/ui/Touch.cpp
@@ -18,7 +18,7 @@ bool Touch::read(int16_t &x, int16_t &y) {
     Wire.beginTransmission(TOUCH_ADDR);
     Wire.write(0x02);                       // Registeranfang: Anzahl Beruehrungen
     if (Wire.endTransmission(false) != 0) return false;
-    if (Wire.requestFrom(TOUCH_ADDR, (uint8_t)5) != 5) return false;
+    if (Wire.requestFrom((uint8_t)TOUCH_ADDR, (uint8_t)5) != 5) return false;
 
     const uint8_t points = Wire.read() & 0x0F;
     const uint8_t xh = Wire.read(), xl = Wire.read();
